Add files via upload
Here are some of my copy and content writing works. My experiences in copy and content writing entail making captions on social platforms, building a company profile for a private business, and writing product descriptions for a booking website. All of it can be seen in the Alfa Arifia Setiawan Portfolio file.